+++ semgrep-core/tests/csharp/dots_nested_stmts.cs
@@ -1,0 +1,8 @@
+public static void Foo()
+{
+    //ERROR:
+    if (x == some_cond)
+        Console.WriteLine("matched");
+    else
+        Console.WriteLine("not matched");
+}