+++ semgrep-core/tests/csharp/dots_string.cs
@@ -1,0 +1,8 @@
+public static void Bar(input) {
+    //ERROR:
+    Foo("whatever sequence of chars");
+
+    //OK:
+    Foo("not a constant string: " + input);
+}
+