+++ semgrep-core/tests/csharp/metavar_call.cs
@@ -1,0 +1,7 @@
+public static void Test
+{
+    // ERROR:
+    Foo(1, 2);
+
+    return 1;
+}