+++ CHANGELOG.md
@@ -6,7 +6,17 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ### Added
 
+### Fixed
+
+### Changed
+
+## [0.52.0](https://github.com/returntocorp/semgrep/releases/tag/v0.52.0) - 2021-05-18
+
+### Added
+
 - C# alpha support
+- Let meta-variables match both a constant variable occurrence and that same
+  constant value (#3058)
 
 ### Fixed
 
@@ -53,8 +63,6 @@ $ARG = [$V];
 ### Added
 - JS/TS: Infer global constants even if the `const` qualifier is missing (#2978)
 - PHP: Resolve names and infer global constants in the same way as for Python
-- Let meta-variables match both a constant variable occurrence and that same
-  constant value
 
 ### Fixed
 - Empty yaml files do not crash